Move absences to own package
