+++ src/main/java/org/synyx/urlaubsverwaltung/absence/AbsenceMappingService.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.calendarintegration.absence;
+package org.synyx.urlaubsverwaltung.absence;
 
 import java.util.Optional;
