+++ src/main/java/org/synyx/urlaubsverwaltung/absence/AbsenceService.java
@@ -1,7 +1,6 @@
 package org.synyx.urlaubsverwaltung.absence;
 
 
-import org.synyx.urlaubsverwaltung.calendarintegration.absence.Absence;
 import org.synyx.urlaubsverwaltung.person.Person;
 
 import java.util.List;
