+++ src/main/java/org/synyx/urlaubsverwaltung/absence/AbsenceTimeConfiguration.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.calendarintegration.absence;
+package org.synyx.urlaubsverwaltung.absence;
 
 import org.synyx.urlaubsverwaltung.settings.CalendarSettings;
