+++ src/main/java/org/synyx/urlaubsverwaltung/absence/AbsenceType.java
@@ -1,0 +1,7 @@
+package org.synyx.urlaubsverwaltung.absence;
+
+public enum AbsenceType {
+
+    VACATION,
+    SICKNOTE
+}
