+++ src/main/java/org/synyx/urlaubsverwaltung/absence/DayAbsence.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.absence.api;
+package org.synyx.urlaubsverwaltung.absence;
 
 import org.synyx.urlaubsverwaltung.api.RestApiDateFormat;
