+++ src/main/java/org/synyx/urlaubsverwaltung/calendar/ICalService.java
@@ -14,7 +14,7 @@ import net.fortuna.ical4j.model.property.XProperty;
 import net.fortuna.ical4j.validate.ValidationException;
 import org.apache.commons.codec.digest.DigestUtils;
 import org.springframework.stereotype.Service;
-import org.synyx.urlaubsverwaltung.calendarintegration.absence.Absence;
+import org.synyx.urlaubsverwaltung.absence.Absence;
 
 import java.time.ZonedDateTime;
 import java.util.List;
