+++ src/main/java/org/synyx/urlaubsverwaltung/calendarintegration/CalendarMailService.java
@@ -1,7 +1,7 @@
 package org.synyx.urlaubsverwaltung.calendarintegration;
 
 import org.springframework.stereotype.Service;
-import org.synyx.urlaubsverwaltung.calendarintegration.absence.Absence;
+import org.synyx.urlaubsverwaltung.absence.Absence;
 import org.synyx.urlaubsverwaltung.mail.MailService;
 
 import java.util.HashMap;
