+++ src/main/java/org/synyx/urlaubsverwaltung/calendarintegration/CalendarSyncServiceImpl.java
@@ -3,7 +3,7 @@ package org.synyx.urlaubsverwaltung.calendarintegration;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.synyx.urlaubsverwaltung.calendarintegration.absence.Absence;
+import org.synyx.urlaubsverwaltung.absence.Absence;
 import org.synyx.urlaubsverwaltung.settings.CalendarSettings;
 import org.synyx.urlaubsverwaltung.settings.SettingsService;
