--- src/main/java/org/synyx/urlaubsverwaltung/calendarintegration/absence/AbsenceType.java
@@ -1,7 +1,0 @@
-package org.synyx.urlaubsverwaltung.calendarintegration.absence;
-
-public enum AbsenceType {
-
-    VACATION,
-    SICKNOTE
-}
