+++ src/main/java/org/synyx/urlaubsverwaltung/calendarintegration/providers/noop/NoopCalendarSyncProvider.java
@@ -2,7 +2,7 @@ package org.synyx.urlaubsverwaltung.calendarintegration.providers.noop;
 
 import org.slf4j.Logger;
 import org.springframework.stereotype.Service;
-import org.synyx.urlaubsverwaltung.calendarintegration.absence.Absence;
+import org.synyx.urlaubsverwaltung.absence.Absence;
 import org.synyx.urlaubsverwaltung.calendarintegration.providers.CalendarProvider;
 import org.synyx.urlaubsverwaltung.settings.CalendarSettings;
