+++ src/test/java/org/synyx/urlaubsverwaltung/calendarintegration/CalendarSyncServiceImplTest.java
@@ -2,7 +2,7 @@ package org.synyx.urlaubsverwaltung.calendarintegration;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.synyx.urlaubsverwaltung.calendarintegration.absence.Absence;
+import org.synyx.urlaubsverwaltung.absence.Absence;
 import org.synyx.urlaubsverwaltung.calendarintegration.providers.exchange.ExchangeCalendarProvider;
 import org.synyx.urlaubsverwaltung.settings.CalendarSettings;
 import org.synyx.urlaubsverwaltung.settings.ExchangeCalendarSettings;
